// client/src/interfaces/IProduct.ts
export interface IProduct {
  id?: string;
  name: string;
  type: string;
  nicotineAmount: string;
  description: string;
  price: number;
  starRating: number;
  imageUrl: string;
  nicotineLevel: string;
  taste: string;
  contentWeight: string;
  contentWeightUnit: string;
  producer: string;
}